add: making sample data.

+++ src/main/java/com/loopy/domain/entity/OrderGroup.java
@@ -34,7 +34,7 @@ public class OrderGroup {
     private String revName;
     private String paymentType;
     private BigDecimal totalPrice;
-    private Long totalQuantity;
+    private int totalQuantity;
     private LocalDateTime orderAt;
     private LocalDateTime arrivalDate;
     @CreatedDate
